+++ PusherServer.Tests/Helpers/EnvironmentVariableConfigLoader.cs
@@ -10,8 +10,7 @@ namespace PusherServer.Tests.Helpers
         private const string PUSHER_APP_ID = "PUSHER_APP_ID";
         private const string PUSHER_APP_KEY = "PUSHER_APP_KEY";
         private const string PUSHER_APP_SECRET = "PUSHER_APP_SECRET";
-        private const string PUSHER_HTTP_HOST = "PUSHER_APP_HOST";
-        private const string PUSHER_WEBSOCKET_HOST = "PUSHER_APP_WEB_SOCKET_HOST";
+        private const string PUSHER_APP_CLUSTER = "PUSHER_APP_CLUSTER";
 
         public static IApplicationConfigLoader Default { get; } = new EnvironmentVariableConfigLoader();
 
@@ -26,8 +25,7 @@ namespace PusherServer.Tests.Helpers
                 AppId = Environment.GetEnvironmentVariable(PUSHER_APP_ID),
                 AppKey = Environment.GetEnvironmentVariable(PUSHER_APP_KEY),
                 AppSecret = Environment.GetEnvironmentVariable(PUSHER_APP_SECRET),
-                HttpHost = Environment.GetEnvironmentVariable(PUSHER_HTTP_HOST),
-                WebSocketHost = Environment.GetEnvironmentVariable(PUSHER_WEBSOCKET_HOST),
+                Cluster = Environment.GetEnvironmentVariable(PUSHER_APP_CLUSTER),
             };
             return result;
         }