+++ PusherServer/Pusher.cs
@@ -1,5 +1,4 @@
 ﻿using System;
-using System.Diagnostics;
 using System.Reflection;
 using System.Threading.Tasks;
 using PusherServer.RestfulClient;
@@ -21,6 +20,8 @@ namespace PusherServer
         private readonly IPusherOptions _options;
 
         private readonly IAuthenticatedRequestFactory _factory;
+
+        private readonly IChannelDataEncrypter _dataEncrypter = new ChannelDataEncrypter();
 
         /// <summary>
         /// Pusher library version information.
@@ -114,22 +115,44 @@ namespace PusherServer
             return result;
         }
 
+        private string SerializeData(string channelName, object data)
+        {
+            string result = _options.JsonSerializer.Serialize(data);
+            if (IsPrivateEncryptedChannel(channelName))
+            {
+                byte[] key = null;
+                if (_options.EncryptionMasterKey != null)
+                {
+                    key = _options.EncryptionMasterKey;
+                }
+
+                EncryptedChannelData encryptedData = _dataEncrypter.EncryptData(channelName, result, key);
+                result = _options.JsonSerializer.Serialize(encryptedData);
+            }
+
+            return result;
+        }
+
         private TriggerBody CreateTriggerBody(string[] channelNames, string eventName, object data, ITriggerOptions options)
         {
             ValidationHelper.ValidateChannelNames(channelNames);
             ValidationHelper.ValidateSocketId(options.SocketId);
 
+            string channelName = null;
+            if (channelNames != null)
+            {
+                if (channelNames.Length > 0)
+                {
+                    channelName = channelNames[0];
+                }
+            }
+
             TriggerBody bodyData = new TriggerBody()
             {
                 name = eventName,
-                data = _options.JsonSerializer.Serialize(data),
+                data = SerializeData(channelName, data),
                 channels = channelNames
             };
-            string channelName = null;
-            if (channelNames != null && channelNames.Length > 0)
-            {
-                channelName = channelNames[0];
-            }
 
             ValidationHelper.ValidateBatchEventData(bodyData.data, channelName, eventName, _options);
 
@@ -154,7 +177,7 @@ namespace PusherServer
                     name = item.EventName,
                     channel = item.Channel,
                     socket_id = item.SocketId,
-                    data = _options.JsonSerializer.Serialize(item.Data),
+                    data = SerializeData(item.Channel, item.Data),
                 };
                 ValidationHelper.ValidateBatchEventData(batchEvent.data, batchEvent.channel, batchEvent.name, _options);
 
@@ -170,7 +193,17 @@ namespace PusherServer
         ///<inheritDoc/>
         public IAuthenticationData Authenticate(string channelName, string socketId)
         {
-            return new AuthenticationData(_appKey, _appSecret, channelName, socketId);
+            IAuthenticationData result;
+            if (IsPrivateEncryptedChannel(channelName))
+            {
+                result = new AuthenticationData(_appKey, _appSecret, channelName, socketId, _options.EncryptionMasterKey);
+            }
+            else
+            {
+                result = new AuthenticationData(_appKey, _appSecret, channelName, socketId);
+            }
+
+            return result;
         }
 
         ///<inheritDoc/>
@@ -232,6 +265,20 @@ namespace PusherServer
             var response = await _options.RestClient.ExecuteGetAsync<T>(request).ConfigureAwait(false);
 
             return response;
+        }
+
+        internal static bool IsPrivateEncryptedChannel(string channelName)
+        {
+            bool result = false;
+            if (channelName != null)
+            {
+                if (channelName.StartsWith("private-encrypted-", StringComparison.OrdinalIgnoreCase))
+                {
+                    result = true;
+                }
+            }
+
+            return result;
         }
 
         private void DebugTriggerRequest(IPusherRestRequest request)