+++ PusherServer/PusherOptions.cs
@@ -87,13 +87,19 @@ namespace PusherServer
             {
                 if (_pusherClient == null)
                 {
-                    _pusherClient = new PusherRestClient(GetBaseUrl(), Pusher.LIBRARY_NAME, Pusher.VERSION);
+                    _pusherClient = new PusherRestClient(GetBaseUrl(), Pusher.LIBRARY_NAME, Pusher.VERSION)
+                    {
+                        Timeout = RestClientTimeout,
+                    };
                 }
 
                 return _pusherClient;
             }
             set { _pusherClient = value; }
         }
+
+        /// <inheritDoc/>
+        public TimeSpan RestClientTimeout { get; set; } = TimeSpan.FromSeconds(30);
 
         /// <inheritDoc/>
         public string HostName
@@ -155,6 +161,9 @@ namespace PusherServer
 
         /// <inheritDoc/>
         public ITraceLogger TraceLogger { get; set; }
+
+        /// <inheritDoc/>
+        public byte[] EncryptionMasterKey { get; set; }
 
         /// <inheritDoc/>
         public Uri GetBaseUrl()