+++ PusherServer/RestfulClient/AuthenticatedRequestFactory.cs
@@ -1,8 +1,6 @@
 ﻿using System;
 using System.Collections.Generic;
 using System.Reflection;
-using System.Text;
-using Newtonsoft.Json;
 
 namespace PusherServer.RestfulClient
 {
@@ -52,9 +50,11 @@ namespace PusherServer.RestfulClient
 
             string requestUrl = $"{path}?auth_signature={authSignature}&{queryString}";
 
-            IPusherRestRequest request = new PusherRestRequest(requestUrl);
-            request.Method = requestType;
-            request.Body = requestBody;
+            IPusherRestRequest request = new PusherRestRequest(requestUrl)
+            {
+                Method = requestType,
+                Body = requestBody,
+            };
 
             return request;
         }